+++ src/QueryCat.Cli/Infrastructure/WebServer.cs
@@ -73,20 +73,21 @@ internal sealed class WebServer
         {
             // Common.
             var context = listener.GetContext();
-            using HttpListenerResponse response = context.Response;
-            response.Headers["User-Agent"] = $"{Application.GetProductFullName()}";
+            var response = context.Response;
+            response.Headers["User-Agent"] = Application.GetProductFullName();
             response.StatusCode = (int)HttpStatusCode.OK;
 
             // CORS.
             if (!string.IsNullOrEmpty(AllowOrigin))
             {
-                context.Response.Headers.Add("Access-Control-Allow-Origin", "*");
+                response.Headers.Add("Access-Control-Allow-Origin", "*");
                 if (context.Request.HttpMethod.Equals(OptionsMethod))
                 {
-                    context.Response.Headers.Add("Access-Control-Allow-Methods", "POST, GET, OPTIONS");
-                    context.Response.Headers.Add("Access-Control-Allow-Headers", "Content-Type");
-                    context.Response.Headers.Add("Access-Control-Max-Age", "86400");
-                    context.Response.StatusCode = (int)HttpStatusCode.OK;
+                    response.Headers.Add("Access-Control-Allow-Methods", "POST, GET, OPTIONS");
+                    response.Headers.Add("Access-Control-Allow-Headers", "Content-Type");
+                    response.Headers.Add("Access-Control-Max-Age", "86400");
+                    response.StatusCode = (int)HttpStatusCode.OK;
+                    response.Close();
                     continue;
                 }
             }
@@ -98,6 +99,7 @@ internal sealed class WebServer
                 if (identity.Password != _password)
                 {
                     response.StatusCode = (int)HttpStatusCode.Unauthorized;
+                    response.Close();
                     continue;
                 }
             }
@@ -112,7 +114,7 @@ internal sealed class WebServer
                 }
                 catch (QueryCatException e)
                 {
-                    using var jsonWriter = new Utf8JsonWriter(context.Response.OutputStream);
+                    using var jsonWriter = new Utf8JsonWriter(response.OutputStream);
                     WriteJsonMessage(jsonWriter, e.Message);
                     response.ContentType = ContentTypeJson;
                     response.StatusCode = (int)HttpStatusCode.BadRequest;
@@ -127,6 +129,8 @@ internal sealed class WebServer
             {
                 response.StatusCode = (int)HttpStatusCode.NotFound;
             }
+
+            response.Close();
         }
         // ReSharper disable once FunctionNeverReturns
     }